# src/main/java/com/example/bankcards/dto/user/UserStatusUpdateRequest.java
package com.example.bankcards.dto.user;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class UserStatusUpdateRequest {
    @NotNull
    private Boolean enabled;
}